The back-off is reachable, and now has the case to prove it (#32)
Review found a concrete input the guard's own comment said did not
exist: `floor_on_grid(800.3, 360.0, 1.7)` computes a line one ulp above
the value it floored from, and without the back-off returns a frame
wider than the space it was asked to fit inside.

My searches missed it because they had the wrong shape — large minima
against tiny steps, on the assumption that the rounding needed extreme
exponents. It needs nothing of the kind. The comment claiming no search
reaches this is gone, and the input is written down.

That also corrects the reasoning I gave for keeping the branch. The
oracle does not backstop it: with the branch deleted the whole suite
stayed green, because the sweep's descriptors never produce the
rounding. A direct case does, and reddens without it.

The oracle's docstring had traded its narrative for a tautology and had
migrated onto the helper below it, leaving the sweep undocumented. Each
now says what its own item does.

The test file crossed its limit, so optimality moved beside soundness
rather than under it: the oracle is a different kind of test — slow,
obviously correct, and the only thing that can tell a suboptimal answer
from a wrong one.
